requirements.txt
# ============================================================================
# ScalpingBot v2.4 Dependencies
# Python 3.11+ Required
# ============================================================================

# -----------------------------------------
# 설정 및 유틸리티
# -----------------------------------------
PyYAML>=6.0.1          # YAML 설정 파일 파싱
python-dotenv>=1.0.0   # 환경 변수 관리

# -----------------------------------------
# HTTP 클라이언트 (한투 API, AI API)
# -----------------------------------------
requests>=2.31.0       # HTTP 요청
aiohttp>=3.9.0         # 비동기 HTTP (선택적)
websockets>=12.0       # 웹소켓 연결 (비동기)
websocket-client>=1.7.0 # 웹소켓 연결 (동기, 실시간 시세용)

# -----------------------------------------
# 웹 스크래핑 (유니버스 구성용)
# -----------------------------------------
beautifulsoup4>=4.12.0 # HTML 파싱

# -----------------------------------------
# 데이터 처리
# -----------------------------------------
pandas>=2.1.0          # 데이터프레임 처리
numpy>=1.26.0          # 수치 계산

# -----------------------------------------
# 데이터베이스
# -----------------------------------------
# SQLite는 Python 내장

# -----------------------------------------
# 스케줄링 및 시간 관리
# -----------------------------------------
schedule>=1.2.0        # 작업 스케줄링
pytz>=2024.1           # 타임존 처리

# -----------------------------------------
# 로깅
# -----------------------------------------
colorlog>=6.8.0        # 컬러 로그 출력

# -----------------------------------------
# 주식 데이터 (유니버스 구성용)
# -----------------------------------------
FinanceDataReader>=0.9.50   # 한국 주식 데이터
# pykrx>=1.0.0          # KRX 데이터 (차단 이슈로 주석)

# -----------------------------------------
# 테스트
# -----------------------------------------
pytest>=7.4.0          # 테스트 프레임워크
pytest-asyncio>=0.21.0 # 비동기 테스트

# -----------------------------------------
# 타입 힌트 및 코드 품질 (개발용)
# -----------------------------------------
mypy>=1.7.0            # 타입 체크
black>=23.12.0         # 코드 포맷터
flake8>=6.1.0          # 린터
